# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaveStream API",
    "version": "0.1.0",
    "description": "Phase 0.5 frozen public contract for SaveStream Web/Mobile. Breaking changes require an explicit contract version change and coordinated frontend/backend update."
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Version 1 base path"
    }
  ],
...
  "paths": {
...
    "/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user",
        "operationId": "getMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user",
        "operationId": "updateMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Request account deletion",
        "operationId": "deleteMe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion workflow accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMessage"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/me/sessions": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List sessions",
        "operationId": "listSessions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/me/sessions/{session_id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Revoke session",
        "operationId": "revokeSession",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "description": "Opaque server-generated identifier. Clients MUST NOT parse it; UUID or ULID may be used internally."
      },
      "UtcDateTime": {
        "type": "string",
        "format": "date-time",
        "description": "UTC ISO-8601 timestamp."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "AUTH_INVALID_CREDENTIALS",
          "AUTH_EMAIL_NOT_VERIFIED",
          "AUTH_SESSION_REVOKED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "VALIDATION_ERROR",
          "RATE_LIMITED",
          "IDEMPOTENCY_KEY_REUSED",
          "STREAM_OFFLINE",
          "STREAM_UNAVAILABLE",
          "RECORDING_ALREADY_ACTIVE",
          "RECORDING_NOT_STOPPABLE",
          "INSUFFICIENT_CREDITS",
          "PAYMENT_PENDING",
          "PAYMENT_FAILED",
          "SERVICE_UNAVAILABLE",
          "INTERNAL_ERROR",
          "PLAN_REQUIRED",
          "WATCH_LIMIT_REACHED",
          "FREE_MINUTES_EXHAUSTED",
          "CREATOR_NOT_LIVE",
          "LOCAL_SLOT_BUSY",
          "EXTENSION_LIMIT_REACHED",
          "REWARD_DAILY_CAP_REACHED",
          "REWARD_LOCKED",
          "STORE_RECEIPT_INVALID",
          "NOT_SOURCE_DEVICE",
          "LOCAL_RECORDING_DISABLED",
          "NOT_IMPLEMENTED"
        ]
      },
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "retryable",
          "details"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "email_verified",
          "display_name",
          "locale",
          "created_at",
          "role",
          "admin_mfa_enabled",
          "admin_mfa_verified"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "email_verified": {
            "type": "boolean"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "locale": {
            "type": "string"
          },
          "created_at": {
            "$ref": "#/components/schemas/UtcDateTime"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "owner",
              "support",
              "finance",
              "admin"
            ]
          },
          "admin_mfa_enabled": {
            "type": "boolean",
            "default": false
          },
          "admin_mfa_verified": {
            "type": "boolean",
            "default": false
          }
        },
        "additionalProperties": false
      },
...
      "AuthMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateMeRequest": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "locale": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Session": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "last_seen_at",
          "current"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "created_at": {
            "$ref": "#/components/schemas/UtcDateTime"
          },
          "last_seen_at": {
            "$ref": "#/components/schemas/UtcDateTime"
          },
          "current": {
            "type": "boolean"
          },
          "user_agent": {
            "type": [
              "string",
              "null"
            ]
          },
          "ip_hint": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "SessionsResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}